% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{age_map}
\alias{age_map}
\title{Age map to divide measurement made in continuous age into age bins}
\format{
A \code{data frame} with columns:
\describe{
\item{label}{Name of age bin}
\item{lo}{lower limit (decimal age)}
\item{hi}{higher limit (decimal age)}
}
}
\usage{
age_map
}
\description{
Age map to divide measurement made in continuous age into age bins
}
\keyword{datasets}
