% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_predictors.R
\name{collect_predictors}
\alias{collect_predictors}
\title{Fill risk predictor data with target person data}
\usage{
collect_predictors(tgt, outcome = "overweight-4y")
}
\arguments{
\item{tgt}{Target person data, according to JAMES internal format from \code{bdsreader}}

\item{outcome}{Character. Currently only \code{"overweight-4y"}}
}
\value{
A tibble with 1 row containing person data
}
\description{
Fill risk predictor data with target person data
}
\examples{
library(bdsreader)
fn <- system.file("examples/test.json", package = "bdsreader")
m <- bdsreader::read_bds(fn)
x <- collect_predictors(m)
}
