% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_varnames.R
\name{map_varnames}
\alias{map_varnames}
\title{Maps variable names from internal to external naming}
\usage{
map_varnames(data = NULL, outcome = "overweight-4y", version = "20230525")
}
\arguments{
\item{data}{Target data, e.g. as produced by \code{collect_predictors()}}

\item{outcome}{Character. Currently only \code{"overweight-4y"}}
}
\value{
A tibble with 1 row containing person data
}
\description{
Use the internal table \code{mapping} to transform the internal data into the
form used by the imported model.
}
\examples{
library(bdsreader)
fn <- system.file("examples/test.json", package = "bdsreader")
m <- bdsreader::read_bds(fn)
x <- collect_predictors(m)
ex <- map_varnames(x)
}
