% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{mapping}
\alias{mapping}
\title{Mapping external - internal variable names}
\format{
A \code{data frame} with columns:
\describe{
\item{outcome}{Outcome name}
\item{version}{Model version}
\item{importance}{Importance score (for random forest/ranger)}
\item{ex_name}{Name used in external model object}
\item{in_name}{Name used internally by bdsmodels}
}
}
\usage{
mapping
}
\description{
Mapping external - internal variable names
}
\keyword{datasets}
