% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc4.R
\docType{data}
\name{pc4}
\alias{pc4}
\title{Postal code covariates}
\format{
\code{pc4}
A data frame with 4052 rows and 3 columns:
\describe{
\item{pc4}{Postal code}
\item{sted}{Urbanicity, 5 levels 1 = low, 5 = high}
\item{woz}{Average WOZ (house) value in postal code}
}
}
\source{
Taken from a file named \code{pc4_2022_v1.xlxs}.
}
\usage{
pc4
}
\description{
Postal codes (4-digits) of The Netherlands with covariates.
}
\keyword{datasets}
